+++ src/main/java/com/groundupworks/android/print/model/Printer.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2014 David Marques.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.groundupworks.android.print.model;
+
+/**
+ * Defines a basic interface with accessor methods for a printer.
+ *
+ * @see com.groundupworks.android.print.GoogleCloudPrint#getPrinters(String)
+ * @see com.groundupworks.android.print.gson.model.GsonPrinter
+ * @see com.groundupworks.android.print.gson.GsonPrinterOperator
+ */
+public interface Printer {
+
+    public static final String ID = "id";
+
+    public static final String NAME = "name";
+
+    public static final String DESCRIPTION = "description";
+
+    public String getId();
+
+    public String getName();
+
+    public String getDescription();
+}
